A robot lost in a maze!
Reeborg was exploring a dark maze and the battery in its flashlight ran out. This was a practice exercise for custom function, while loops and Boolean statements.
My program helps Reeborg find the exit. He follows right edge of the maze, turning right if it can, going straight ahead if it can’t turn right, or turning left as a last resort.